backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
requests==2.31.0
python-dotenv==1.0.0
